Aggregating observations on some variables

aggregate_by aggregates some variables of a data frame according to the values of a categorical variable of the same data frame and using a given function for the aggregation.

aggregate_by(df, col_name, ..., .funs = sum)

Arguments

df
the data frame on which to perform the aggregation.

col_name
the variable on which to perform the aggregation

...
other variables on which to perform the aggregation

.funs
the function used to perform the aggregation. By default sum.

Value

A data frame with the same variables as `df` but for which some of the observation have been aggregated (i.e. less rows than in `df`).

Details

On a data frame that contains the values of the categorical variable to aggregate (`col_name`, `...`), the function performs a group_by followed by a summarise with the function(s) .funs.

Examples

library(dplyr)

## A toy data frame:
set.seed(30101976)
data <- expand.grid(letters[1:3], 1:3, 4:6)
data$Var4 <- sample(1:100, nrow(data), TRUE)
data$Var5 <- sample(1:100, nrow(data), TRUE)
data$Var6 <- sample(1:100, nrow(data), TRUE)

## Aggregating the values "a" and "b" of the categorical variable Var1,
## summing the values of variables Var4, Var5, Var6 (i.e. all the variables
## that are not in the arguments of the function call):
data %>%
 mutate(Var1 = recode(Var1, a = "b")) %>%
 aggregate_by(data, Var1, Var2, Var3)

#> Warning: Truncating vector to length 1

#> Warning: Truncating vector to length 1

#> Warning: Truncating vector to length 1

#> Warning: Truncating vector to length 1

#> Warning: Truncating vector to length 1

#> Warning: Truncating vector to length 1

#> [90m# A tibble: 1 x 6[39m
#> [90m# Groups:   Var1, Var2, Var3, Var4, Var5 [?][39m
#>   Var1   Var2  Var3  Var4  Var5  Var6
#>   [3m[90m<fct>[39m[23m [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m
#> [90m1[39m a         1     4    58    67    66