Aggregating observations on some variables

aggregate_by aggregates some variables of a data frame according to the values of a categorical variable of the same data frame and using a given function for the aggregation.

aggregate_by(df, what, ..., fun = sum, new_name = NULL)

Arguments

df
the data frame on which to perform the aggregation.

what
an expression on the values of the categorical variable used for aggregation.

...
the variables (unquoted names) on which to perform the aggregation, using the funtion fun.

fun
the function used to perform the aggregation. By default sum.

new_name
the new value of the categorical variable, after aggregation. By default, it concatenates the values of the categorical variables used in what, separating them by " & ".

Value

A data frame with the same variables as `df` but for which some of the observation have been aggregated (i.e. less rows than in `df`).

Details

The function splits the data frame df in two data frames, based on the what argument. On the data frame that contains the values of the categorical variable to aggregate (df1), the function performs a group_by followed by a summarise_all with the function fun and then binds it by row with the other data frame (df2).

Examples

## A toy data frame:
set.seed(30101976)
data <- expand.grid(letters[1:3], 1:3, 4:6)
data$Var4 <- sample(1:100, nrow(data), TRUE)
data$Var5 <- sample(1:100, nrow(data), TRUE)
data$Var6 <- sample(1:100, nrow(data), TRUE)

## Aggregating the values "a" and "b" of the categorical variable Var1,
## summing the values of variables Var4, Var5, Var6 (i.e. all the variables
## that are not in the arguments of the function call):
aggregate_by(data, Var1 %in% c("a", "b"), Var2, Var3)

#>     Var1 Var2 Var3 Var4 Var5 Var6
#> 1      c    1    4   49   12   62
#> 2      c    2    4   42    9    6
#> 3      c    3    4   16    2   96
#> 4      c    1    5    5   27   17
#> 5      c    2    5   27   35   52
#> 6      c    3    5   21    5   10
#> 7      c    1    6   41   10   43
#> 8      c    2    6   35   38    3
#> 9      c    3    6   28   27    7
#> 10 a & b    1    4   77  136  142
#> 11 a & b    1    5  119  116  106
#> 12 a & b    1    6   14  111  107
#> 13 a & b    2    4   77  122  117
#> 14 a & b    2    5   78   33   98
#> 15 a & b    2    6  113  144  171
#> 16 a & b    3    4   54   81   82
#> 17 a & b    3    5  139  138  123
#> 18 a & b    3    6   49   47   58